pyproject.toml
[project]
name = "civic-data-boundaries-us-forests"
version = "0.0.1"
description = "Geographic boundary layers for Civic Interconnect projects"
authors = [
    { name="Civic Interconnect", email="info@civic-interconnect.org" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "PyYAML",
    "pandas",
    "shapely",
    "pyproj",
    "geopandas",
    "pyogrio",
    "typer[all]",
    "rich",
    "civic-lib-core @ git+https://github.com/civic-interconnect/civic-lib-core.git@main",
    "civic-lib-geo @ git+https://github.com/civic-interconnect/civic-lib-geo.git@main",
    "certifi",
    "python-certifi-win32"
]

[project.optional-dependencies]
dev = [
  "mkdocs",
  "mkdocs-material",
  "pre-commit",
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
  "ruff",
]

[project.urls]
"Homepage" = "https://github.com/civic-interconnect/civic-data-boundaries-us-forests"
"Source" = "https://github.com/civic-interconnect/civic-data-boundaries-us-forests"

[project.scripts]
civic-dev = "civic_dev.cli:app"
civic-usa = "civic_data_boundaries_us_forests.cli.cli:app"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.ruff]
line-length = 100
target-version = "py312"
fix = true
unsafe-fixes = true
preview = true

[tool.ruff.lint]
extend-select = ["E", "F", "W", "I", "UP", "C90", "N", "B", "A", "SIM", "TCH"]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.format]
docstring-code-format = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra --strict-markers"
pythonpath = ["src"]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = true
